# api/swagger/swagger.yaml
swagger: "2.0"
info:
  version: "0.0.1"
  title: Hello World App
# during dev, should point to your local machine
host: localhost:10010
# basePath prefixes all resource paths
basePath: /
#
schemes:
  # tip: remove http to make production-grade
  - http
  - https
# format of bodies a client can send (Content-Type)
consumes:
  - application/json
# format of the responses to the client (Accepts)
produces:
  - application/json
paths:
  /hello:
    # binds a127 app logic to a route
    x-swagger-router-controller: hello_world
    get:
      description: Returns 'Hello' to the caller
      # used as the method name of the controller
      operationId: hello
      parameters:
        - name: name
          in: query
          description: The name of the person to whom to say hello
          required: false
          type: string
      responses:
        "200":
          description: Success
          schema:
            # a pointer to a definition
            $ref: "#/definitions/HelloWorldResponse"
        # responses may fall through to errors
        default:
          description: Error
          schema:
            $ref: "#/definitions/ErrorResponse"
  /iot:
    # binds a127 app logic to a route
    x-swagger-router-controller: iotProj
    post:
      description: For creating table and insert values.
      # used as the method name of the controller
      operationId: iotOperation
      parameters:
        - name: name
          in: body
          description: The name of the person to whom to say hello
          schema:
            # a pointer to a definition
            $ref: "#/definitions/iotRequest"
      responses:
        "200":
          description: Success
          schema:
            # a pointer to a definition
            $ref: "#/definitions/iotResponse"
        # responses may fall through to errors
        default:
          description: Error
          schema:
            $ref: "#/definitions/ErrorResponse"
  /swagger:
    x-swagger-pipe: swagger_raw
# complex objects have schema definitions
definitions:
  iotRequest:
    required:
      - tableName  
      - col1
      - col2
      - col3
      - val1
      - val2
      - val3
    properties:
      tableName:
        type: string
      col1:
        type: string
      col2:
        type: string
      col3:
        type: string
      val1:
        type: number
      val2:
        type: string
      val3:
        type: string
  iotResponse:
    required:
      - message
    properties:
      message:
        type: string
  HelloWorldResponse:
    required:
      - message
    properties:
      message:
        type: string
  ErrorResponse:
    required:
      - message
    properties:
      message:
        type: string